Better separation between parsers that parse whole files and those who parse only certain structures
