+++ src/Parser.cpp
@@ -32,6 +32,8 @@
 #include "Files/DsnStream.hpp"
 #include "Files/HSObjects.hpp"
 #include "Files/NetBundleMapData.hpp"
+#include "Files/SymbolsLibrary.hpp"
+#include "Files/Type.hpp"
 #include "General.hpp"
 #include "Parser.hpp"
 #include "Parser.hpp"
@@ -45,9 +47,7 @@
 #include "Structures/SymbolDisplayProp.hpp"
 #include "Structures/SymbolPinBus.hpp"
 #include "Structures/SymbolPinScalar.hpp"
-#include "Structures/SymbolsLibrary.hpp"
 #include "Structures/T0x1f.hpp"
-#include "Structures/Type.hpp"
 
 
 namespace fs = std::filesystem;
@@ -215,7 +215,8 @@ Library Parser::parseLibrary()
         }
         else
         {
-            throw std::runtime_error("Did not expect to find files here!");
+            throw std::runtime_error(
+                fmt::format("{}: Did not expect to find files here!", __func__));
         }
     }
 
@@ -362,7 +363,7 @@ Library Parser::parseLibrary()
 
     if(fs::exists(pathGraphicsTypes))
     {
-        mLibrary.graphicsTypes = parseFile<std::vector<Type>>(pathGraphicsTypes, [this](){ return parseTypes(); });
+        mLibrary.graphicsTypes = parseFile<std::vector<Type>>(pathGraphicsTypes, [this](){ return readType(); });
     }
     else
     {
@@ -373,7 +374,7 @@ Library Parser::parseLibrary()
 
     if(fs::exists(pathSymbolsTypes))
     {
-        mLibrary.symbolsTypes = parseFile<std::vector<Type>>(pathSymbolsTypes, [this](){ return parseTypes(); });
+        mLibrary.symbolsTypes = parseFile<std::vector<Type>>(pathSymbolsTypes, [this](){ return readType(); });
     }
     else
     {
@@ -384,7 +385,7 @@ Library Parser::parseLibrary()
 
     if(fs::exists(pathLibrary))
     {
-        mLibrary.symbolsLibrary = parseFile<SymbolsLibrary>(pathLibrary, [this](){ return parseSymbolsLibrary(); });
+        mLibrary.symbolsLibrary = parseFile<SymbolsLibrary>(pathLibrary, [this](){ return readSymbolsLibrary(); });
     }
     else
     {
@@ -418,7 +419,7 @@ Library Parser::parseLibrary()
         {
             const fs::path& pathPackage = file.path();
 
-            mLibrary.packages.push_back(parseFile<Package>(pathPackage, [this](){ return parsePackage(); }));
+            mLibrary.packages.push_back(parseFile<Package>(pathPackage, [this](){ return readPackage(); }));
 
             spdlog::info("----------------------------------------------------------------------------------\n");
         }
@@ -444,7 +445,7 @@ Library Parser::parseLibrary()
 
             // @todo Results are only stored in packages for testing purposes
             //       Replace with symbols later on.
-            mLibrary.packages.push_back(parseFile<Package>(pathSymbol, [this](){ return parseSymbol(); }));
+            mLibrary.packages.push_back(parseFile<Package>(pathSymbol, [this](){ return readSymbol(); }));
 
             spdlog::info("----------------------------------------------------------------------------------\n");
         }
@@ -459,7 +460,7 @@ Library Parser::parseLibrary()
         if(fs::exists(schematic))
         {
             // @todo write results into mLibrary
-            parseFile<bool>(schematic, [this](){ return parseSchematic(); });
+            parseFile<bool>(schematic, [this](){ return readSchematic(); });
         }
         else
         {
@@ -474,7 +475,7 @@ Library Parser::parseLibrary()
         if(fs::exists(hierarchy))
         {
             // @todo write results into mLibrary
-            parseFile<bool>(hierarchy, [this](){ return parseHierarchy(); });
+            parseFile<bool>(hierarchy, [this](){ return readHierarchy(); });
         }
         else
         {
@@ -491,7 +492,7 @@ Library Parser::parseLibrary()
             if(fs::exists(page))
             {
                 // @todo write results into mLibrary
-                parseFile<bool>(page, [this](){ return parsePage(); });
+                parseFile<bool>(page, [this](){ return readPage(); });
             }
             else
             {
@@ -535,225 +536,6 @@ void Parser::exceptionHandling()
         spdlog::error(fmt::format(fg(fmt::color::crimson), "--------ERROR REPORT--------"));
         spdlog::error(fmt::format(fg(fmt::color::crimson), "Unknown exception caught!\n"));
     }
-}
-
-
-// @todo return real data object
-bool Parser::parsePage()
-{
-    spdlog::debug(getOpeningMsg(__func__, mDs.getCurrentOffset()));
-
-    bool obj = false;
-
-    // readDevHelper();
-    // return;
-
-    mDs.printUnknownData(21, std::string(__func__) + " - 0");
-    readPreamble();
-
-    std::string name = mDs.readStringLenZeroTerm();
-
-    std::string pageSize = mDs.readStringLenZeroTerm();
-
-    time_t createDateTime = static_cast<time_t>(mDs.readUint32());
-    time_t modifyDateTime = static_cast<time_t>(mDs.readUint32());
-
-    spdlog::debug("createDateTime = {} | modifyDateTime = {}", DateTimeToStr(createDateTime), DateTimeToStr(modifyDateTime));
-
-    mDs.printUnknownData(16, std::string(__func__) + " - 1");
-
-    uint32_t width  = mDs.readUint32();
-    uint32_t height = mDs.readUint32();
-
-    spdlog::debug("width = {} | height = {}", width, height);
-
-    uint32_t pinToPin = mDs.readUint32(); //!< Pin-to-pin spacing
-
-    spdlog::debug("pinToPin = {}", pinToPin);
-
-    mDs.printUnknownData(2, std::string(__func__) + " - 2");
-
-    uint16_t horizontalCount = mDs.readUint16(); //!< See 'Schematic Page Properties' -> 'Grid Reference'
-    uint16_t verticalCount   = mDs.readUint16(); //!< See 'Schematic Page Properties' -> 'Grid Reference'
-
-    spdlog::debug("horizontalCount = {} | verticalCount = {}", horizontalCount, verticalCount);
-
-    mDs.printUnknownData(2, std::string(__func__) + " - 3");
-
-    uint32_t horizontalWidth = mDs.readUint32(); //!< See 'Schematic Page Properties' -> 'Grid Reference'
-    uint32_t verticalWidth   = mDs.readUint32(); //!< See 'Schematic Page Properties' -> 'Grid Reference'
-
-    spdlog::debug("horizontalWidth = {} | verticalWidth = {}", horizontalWidth, verticalWidth);
-
-    mDs.printUnknownData(48, std::string(__func__) + " - 4");
-
-    uint32_t horizontalChar = mDs.readUint32(); //!<       See 'Schematic Page Properties' -> 'Grid Reference' ->
-                                             //             'Horizontal' -> 'Alphabetic' = 1
-                                             //                             'Numeric'    = 0
-
-    spdlog::debug("horizontalChar = {}", horizontalChar);
-
-    mDs.printUnknownData(4, std::string(__func__) + " - 5");
-
-    uint32_t horizontalAscending = mDs.readUint32(); //!<       See 'Schematic Page Properties' -> 'Grid Reference' ->
-                                                  //         'Horizontal'
-                                                  //         Select 'ascending' (1) or 'decending' (0).
-
-    spdlog::debug("horizontalAscending = {}", horizontalAscending);
-
-    uint32_t verticalChar   = mDs.readUint32(); //!< See 'Schematic Page Properties' -> 'Grid Reference' ->
-                                             //       'Vertical' -> 'Alphabetic' = 1
-                                             //                     'Numeric'    = 0
-
-    spdlog::debug("verticalChar = {}", verticalChar);
-
-    mDs.printUnknownData(4, std::string(__func__) + " - 6");
-
-
-    uint32_t verticalAscending = mDs.readUint32(); //!<       See 'Schematic Page Properties' -> 'Grid Reference' ->
-                                                  //         'Vertical'
-                                                  //         Select 'ascending' (1) or 'decending' (0).
-
-    spdlog::debug("verticalAscending = {}", verticalAscending);
-
-    uint32_t isMetric            = mDs.readUint32();
-    uint32_t borderDisplayed     = mDs.readUint32();
-    uint32_t borderPrinted       = mDs.readUint32();
-    uint32_t gridRefDisplayed    = mDs.readUint32();
-    uint32_t gridRefPrinted      = mDs.readUint32();
-    uint32_t titleblockDisplayed = mDs.readUint32();
-    uint32_t titleblockPrinted   = mDs.readUint32();
-    // @todo XML is always 1, maybe a bug in OrCad?
-    uint32_t ansiGridRefs = mDs.readUint32(); //!< Use ANSI grid references.
-                                              //   See 'Schematic Page Properties' -> 'Grid Reference'
-
-    spdlog::debug("isMetric            = {}", isMetric);
-    spdlog::debug("borderDisplayed     = {}", borderDisplayed);
-    spdlog::debug("borderPrinted       = {}", borderPrinted);
-    spdlog::debug("gridRefDisplayed    = {}", gridRefDisplayed);
-    spdlog::debug("gridRefPrinted      = {}", gridRefPrinted);
-    spdlog::debug("titleblockDisplayed = {}", titleblockDisplayed);
-    spdlog::debug("titleblockPrinted   = {}", titleblockPrinted);
-    spdlog::debug("ansiGridRefs        = {}", ansiGridRefs);
-
-    const uint16_t lenA = mDs.readUint16();
-
-    spdlog::debug("lenA = {}", lenA);
-
-    for(size_t i = 0u; i < lenA; ++i)
-    {
-        mDs.printUnknownData(8, std::string(__func__) + " - a");
-    }
-
-    const uint16_t len0 = mDs.readUint16();
-
-    spdlog::debug("len0 = {}", len0);
-
-    for(size_t i = 0u; i < len0; ++i)
-    {
-        mDs.printUnknownData(32, std::string(__func__) + " - 8");
-    }
-
-    // mDs.printUnknownData(30, std::string(__func__) + " - 7");
-
-
-    // mDs.printUnknownData(38, std::string(__func__) + " - 1.5");
-
-    // @todo required for CONTENT page but not for the others? This offset must be somehow
-    //       dynamic
-    // mDs.printUnknownData(14, std::string(__func__) + " - 1.6");
-
-    mDs.printUnknownData(2, std::string(__func__) + " - 9");
-
-    const uint16_t len1 = mDs.readUint16();
-
-    spdlog::debug("len1 = {}", len1);
-
-    for(size_t i = 0u; i < len1; ++i)
-    {
-        std::string name = mDs.readStringLenZeroTerm();
-        mDs.printUnknownData(4, std::string(__func__) + " - 10");
-    }
-
-    const uint16_t len2 = mDs.readUint16();
-
-    spdlog::debug("len2 = {}", len2);
-
-    for(size_t i = 0u; i < len2; ++i)
-    {
-        // if(i > 0u)
-        // {
-        //     readPreamble();
-        //     mDs.printUnknownData(8, std::string(__func__) + " - 11");
-        // }
-
-        Structure structure = read_type_prefix();
-        readPreamble();
-        parseStructure(structure); // @todo push structure
-
-        // structure = read_type_prefix();
-        // readPreamble();
-        // parseStructure(structure); // @todo push structure
-    }
-
-    const uint16_t len3 = mDs.readUint16();
-
-    spdlog::debug("len3 = {}", len3);
-
-    for(size_t i = 0u; i < len3; ++i)
-    {
-        // if(i > 0u)
-        // {
-        //     readPreamble();
-        //     mDs.printUnknownData(8, std::string(__func__) + " - 11");
-        // }
-
-        Structure structure;
-
-        if(i == 0u)
-        {
-            // @todo this is type_prefix_very_long()
-            mDs.printUnknownData(47, std::string(__func__) + " - 11");
-            structure = ToStructure(0x0d); // Parse package instance for now until type_prefix_very_long is implemented
-        }
-        else
-        {
-            structure = read_type_prefix();
-        }
-
-
-        readPreamble();
-        parseStructure(structure); // @todo push structure
-
-        // structure = read_type_prefix();
-        // readPreamble();
-        // parseStructure(structure); // @todo push structure
-    }
-
-    // readPreamble();
-    // mDs.printUnknownData(12, std::string(__func__) + " - 12");
-
-    mDs.printUnknownData(10, std::string(__func__) + " - 10");
-
-    const uint16_t lenX = mDs.readUint16();
-
-    spdlog::debug("lenX = {}", lenX);
-
-    for(size_t i = 0u; i < lenX; ++i)
-    {
-        Structure structure = read_type_prefix();
-        readPreamble();
-        parseStructure(structure); // @todo push structure
-    }
-
-    if(!mDs.isEoF())
-    {
-        throw std::runtime_error("Expected EoF but did not reach it!");
-    }
-
-    spdlog::debug(getClosingMsg(__func__, mDs.getCurrentOffset()));
-
-    return obj;
 }
 
 
@@ -1794,40 +1576,6 @@ Properties2 Parser::readProperties2()
     spdlog::info(to_string(obj));
 
     return obj;
-}
-
-
-std::vector<Type> Parser::parseTypes()
-{
-    spdlog::debug(getOpeningMsg(__func__, mDs.getCurrentOffset()));
-
-    std::vector<Type> types;
-
-    // File can be completely empty (size of 0 byte)
-    while(!mDs.isEoF())
-    {
-        Type type;
-
-        type.name = mDs.readStringLenZeroTerm();
-        type.componentType = ToComponentType(mDs.readUint16());
-
-        types.push_back(type);
-    }
-
-    spdlog::debug(getClosingMsg(__func__, mDs.getCurrentOffset()));
-
-    // @todo move to separate to_string method
-    std::string str;
-
-    for(size_t i = 0U; i < types.size(); ++i)
-    {
-        str += fmt::format("[{:>3}]:\n", i);
-        str += indent(to_string(types[i]), 2);
-    }
-
-    spdlog::info(str);
-
-    return types;
 }
